requirements.txt
openai
chromadb
python-dotenv
tiktoken
requests
beautifulsoup4
langchain-text-splitters
numpy

